fix: try tests
- Introduce unit tests for `MovieDao` to verify database operations.
- Refactor `MovieEntity` to correctly implement `equals` and `hashCode` and adjust runtime and voteAverage types.
- Update `TempUtil` movie mapping functions to align with the `MovieEntity` changes.
- Add androidx.junit-ktx dependency and configure Android unit tests to include resources.

+++ app/src/main/java/com/eaccid/musimpa/data/domain/Movie.kt
@@ -11,7 +11,7 @@ data class Movie(
     val posterPath: String? = null,
     val title: String? = null,
     val overview: String? = null,
-    val voteAverage: Int = 0, //percentage
+    val voteAverage: Double = 0.0, //percentage
     val tagline: String? = null,
     val runtime: Int = 0,
     var videoKey: String = ""
